changed structure and fixed featurebuilder type hints
