+++ text_authorship/ta_model/__init__.py
@@ -2,4 +2,4 @@ from .data_extraction import load_df # noqa
 from .model_selection import train_crossval_twofold # noqa
 from .model_selection import train_test_split # noqa
 from .model_selection import get_encoders # noqa
-from featurebuilder import FeatureBuilder # noqa
+from vectorizers.featurebuilder import FeatureBuilder # noqa
