+++ text_authorship/ta_model/base_models.py
@@ -6,9 +6,9 @@ from xgboost import XGBClassifier
 from config.config_parser import parse_config
 
 from text_authorship.ta_model import train_test_split, get_encoders
-from text_authorship.ta_model.data_preparation import get_encoder
+from data_preparation.features import get_encoder
 from stacking.tastack2 import TAStack2
-from stacking.tavectorizer import TAVectorizer
+from vectorizers.tavectorizer import TAVectorizer
 
 _FEATURE_PARAMS = "feature_params"
