+++ text_authorship/ta_model/parsing/parsers/wordparser.py
@@ -1,19 +1,22 @@
 from constants import ParsingParams, TokenType, DELETED
-from parsemanager import ParseManager
-from parsers.baseparser import BaseParser
+from parsing.parsemanager import ParseManager
+from parsing.parsers.baseparser import BaseParser
 
 
 @ParseManager.register_parser
-class TokenParser(BaseParser):
+class WordParser(BaseParser):
 
-    col_name = 'tokens'
+    col_name = 'text_no_punkt'
 
     def __init__(self):
         super().__init__()
 
     def parse_token(self, params: ParsingParams) -> None:
+        if TokenType.PUNKT in params.token_type:
+            return
+
         if TokenType.STOPTAG in params.token_type:
             self.tokens.append(DELETED)
             return
 
-        self.tokens.append(params.token)
+        self.tokens.append(params.stripped)
